+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/feed/models/FeedAction.kt
@@ -1,5 +1,5 @@
 package com.mobiledeveloper.vktube.ui.screens.feed.models
 
 sealed class FeedAction {
-    data class OpenVideoDetail(val videoId: Long) : FeedAction()
+    data class OpenVideoDetail(val videoId: Int) : FeedAction()
 }
