+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/video/VideoScreen.kt
@@ -44,7 +44,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalFoundationApi::class, ExperimentalMaterialApi::class)
 @Composable
 fun VideoScreen(
-    videoId: Long?,
+    videoId: Int?,
     videoViewModel: VideoViewModel
 ) {
     val viewState by videoViewModel.viewStates().collectAsState()
