+++ app/src/main/java/com/mobiledeveloper/vktube/ui/screens/video/models/VideoEvent.kt
@@ -1,7 +1,7 @@
 package com.mobiledeveloper.vktube.ui.screens.video.models
 
 sealed class VideoEvent {
-    data class LaunchVideo(val videoId: Long?) : VideoEvent()
+    data class LaunchVideo(val videoId: Int?) : VideoEvent()
     data class SendComment(val comment: String) : VideoEvent()
     object LikeClick : VideoEvent()
     object CommentsClick : VideoEvent()
